consolidate creatorgetters
